+++ mongodbatlas/fw_resource_mongodbatlas_project_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"regexp"
 	"strings"
 	"testing"
 
@@ -15,51 +16,31 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
-// This test only helps to see what the plan is for debugging purposes
-// func Test_DebugPlan(t *testing.T) {
-// 	var (
-// 		// resourceName = "mongodbatlas_project.test"
-// 		projectName = acctest.RandomWithPrefix("test-acc-migration")
-// 		orgID       = os.Getenv("MONGODB_ATLAS_ORG_ID")
-// 	)
-// 	t.Parallel()
+type projectLimit struct {
+	name  string
+	value int64
+}
 
-// 	resource.Test(t, resource.TestCase{
-// 		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
-// 		Steps: []resource.TestStep{
-// 			{
-// 				Config: fmt.Sprintf(`resource "mongodbatlas_project" "main" {
-// 					name   = "%s"
-// 					org_id = "%s"
-// 				  }`, projectName, orgID),
-// 				PlanOnly: true,
-// 				ConfigPlanChecks: resource.ConfigPlanChecks{
-// 					PostApplyPreRefresh: []plancheck.PlanCheck{
-// 						DebugPlan(),
-// 					},
-// 				},
-// 			},
-// 		},
-// 	})
-// }
-
-func TestAccRSProject_Teams(t *testing.T) {
+func TestAccProjectRSProject_basic(t *testing.T) {
 	var (
 		project      matlas.Project
 		resourceName = "mongodbatlas_project.test"
-		projectName  = acctest.RandomWithPrefix("test-acc-teams")
+		projectName  = acctest.RandomWithPrefix("test-acc")
 		orgID        = os.Getenv("MONGODB_ATLAS_ORG_ID")
 		clusterCount = "0"
 		teamsIds     = strings.Split(os.Getenv("MONGODB_ATLAS_TEAMS_IDS"), ",")
 	)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheckBasic(t); testCheckTeamsIds(t) },
-		CheckDestroy: testAccCheckMongoDBAtlasProjectDestroy2,
+	if len(teamsIds) < 3 {
+		t.Skip("`MONGODB_ATLAS_TEAMS_IDS` must have 3 team ids for this acceptance testing")
+	}
 
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t); testCheckTeamsIds(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
 		Steps: []resource.TestStep{
 			{
-				ProtoV6ProviderFactories: testProtoV6ProviderFactories,
-				Config: testAccMongoDBAtlasProjectConfig2(projectName, orgID,
+				Config: testAccMongoDBAtlasProjectConfig(projectName, orgID,
 					[]*matlas.ProjectTeam{
 						{
 							TeamID:    teamsIds[0],
@@ -69,10 +50,69 @@ func TestAccRSProject_Teams(t *testing.T) {
 							TeamID:    teamsIds[1],
 							RoleNames: []string{"GROUP_DATA_ACCESS_ADMIN", "GROUP_OWNER"},
 						},
-					}),
+					},
+				),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckMongoDBAtlasProjectExists2(resourceName, &project),
-					testAccCheckMongoDBAtlasProjectAttributes2(&project, projectName),
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfig(projectName, orgID,
+
+					[]*matlas.ProjectTeam{
+						{
+							TeamID:    teamsIds[0],
+							RoleNames: []string{"GROUP_OWNER"},
+						},
+						{
+							TeamID:    teamsIds[1],
+							RoleNames: []string{"GROUP_DATA_ACCESS_READ_WRITE"},
+						},
+						{
+							TeamID:    teamsIds[2],
+							RoleNames: []string{"GROUP_READ_ONLY", "GROUP_DATA_ACCESS_ADMIN"},
+						},
+					},
+				),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfig(projectName, orgID,
+
+					[]*matlas.ProjectTeam{
+						{
+							TeamID:    teamsIds[0],
+							RoleNames: []string{"GROUP_READ_ONLY", "GROUP_READ_ONLY"},
+						},
+						{
+							TeamID:    teamsIds[1],
+							RoleNames: []string{"GROUP_OWNER", "GROUP_DATA_ACCESS_ADMIN"},
+						},
+					},
+				),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfig(projectName, orgID, []*matlas.ProjectTeam{}),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
 					resource.TestCheckResourceAttr(resourceName, "name", projectName),
 					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
 					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
@@ -82,46 +122,285 @@ func TestAccRSProject_Teams(t *testing.T) {
 	})
 }
 
-func testAccCheckMongoDBAtlasProjectDestroy2(s *terraform.State) error {
-	conn := testMongoDBClient.(*MongoDBClient).Atlas
+func TestAccProjectRSProject_CreateWithProjectOwner(t *testing.T) {
+	var (
+		project        matlas.Project
+		resourceName   = "mongodbatlas_project.test"
+		projectName    = acctest.RandomWithPrefix("test-acc")
+		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID")
+		projectOwnerID = os.Getenv("MONGODB_ATLAS_PROJECT_OWNER_ID")
+	)
 
-	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "mongodbatlas_project" {
-			continue
-		}
-
-		projectRes, _, _ := conn.Projects.GetOneProjectByName(context.Background(), rs.Primary.ID)
-		if projectRes != nil {
-			return fmt.Errorf("project (%s) still exists", rs.Primary.ID)
-		}
-	}
-
-	return nil
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasicOwnerID(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithProjectOwner(projectName, orgID, projectOwnerID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+				),
+			},
+		},
+	})
 }
 
-func testAccMongoDBAtlasProjectConfig2(projectName, orgID string, teams []*matlas.ProjectTeam) string {
-	var ts string
+func TestAccProjectRSGovProject_CreateWithProjectOwner(t *testing.T) {
+	var (
+		project        matlas.Project
+		resourceName   = "mongodbatlas_project.test"
+		projectName    = acctest.RandomWithPrefix("tf-acc-project")
+		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID_GOV")
+		projectOwnerID = os.Getenv("MONGODB_ATLAS_PROJECT_OWNER_ID_GOV")
+	)
 
-	for _, t := range teams {
-		ts += fmt.Sprintf(`
-		teams {
-			team_id = "%s"
-			role_names = %s
-		}
-		`, t.TeamID, strings.ReplaceAll(fmt.Sprintf("%+q", t.RoleNames), " ", ","))
-	}
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckGov(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
 
-	return fmt.Sprintf(`
-		resource "mongodbatlas_project" "test" {
-			name  			 = "%s"
-			org_id 			 = "%s"
+		CheckDestroy: testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasGovProjectConfigWithProjectOwner(projectName, orgID, projectOwnerID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+				),
+			},
+		},
+	})
+}
+func TestAccProjectRSProject_CreateWithFalseDefaultSettings(t *testing.T) {
+	var (
+		project        matlas.Project
+		resourceName   = "mongodbatlas_project.test"
+		projectName    = acctest.RandomWithPrefix("tf-acc-project")
+		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID")
+		projectOwnerID = os.Getenv("MONGODB_ATLAS_PROJECT_OWNER_ID")
+	)
 
-			%s
-		}
-	`, projectName, orgID, ts)
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasicOwnerID(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithFalseDefaultSettings(projectName, orgID, projectOwnerID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+				),
+			},
+		},
+	})
 }
 
-func testAccCheckMongoDBAtlasProjectExists2(resourceName string, project *matlas.Project) resource.TestCheckFunc {
+func TestAccProjectRSProject_CreateWithFalseDefaultAdvSettings(t *testing.T) {
+	var (
+		project        matlas.Project
+		resourceName   = "mongodbatlas_project.test"
+		projectName    = acctest.RandomWithPrefix("tf-acc-project")
+		orgID          = os.Getenv("MONGODB_ATLAS_ORG_ID")
+		projectOwnerID = os.Getenv("MONGODB_ATLAS_PROJECT_OWNER_ID")
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasicOwnerID(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithFalseDefaultAdvSettings(projectName, orgID, projectOwnerID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckMongoDBAtlasProjectExists(resourceName, &project),
+					testAccCheckMongoDBAtlasProjectAttributes(&project, projectName),
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+				),
+			},
+		},
+	})
+}
+
+func TestAccProjectRSProject_withUpdatedRole(t *testing.T) {
+	var (
+		resourceName    = "mongodbatlas_project.test"
+		projectName     = acctest.RandomWithPrefix("tf-acc-project")
+		orgID           = os.Getenv("MONGODB_ATLAS_ORG_ID")
+		roleName        = "GROUP_DATA_ACCESS_ADMIN"
+		roleNameUpdated = "GROUP_READ_ONLY"
+		clusterCount    = "0"
+		teamsIds        = strings.Split(os.Getenv("MONGODB_ATLAS_TEAMS_IDS"), ",")
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t); testCheckTeamsIds(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithUpdatedRole(projectName, orgID, teamsIds[0], roleName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithUpdatedRole(projectName, orgID, teamsIds[0], roleNameUpdated),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "cluster_count", clusterCount),
+				),
+			},
+		},
+	})
+}
+
+func TestAccProjectRSProject_importBasic(t *testing.T) {
+	var (
+		projectName  = acctest.RandomWithPrefix("tf-acc-project")
+		orgID        = os.Getenv("MONGODB_ATLAS_ORG_ID")
+		resourceName = "mongodbatlas_project.test"
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfig(projectName, orgID,
+					[]*matlas.ProjectTeam{},
+				),
+			},
+			{
+				ResourceName:            resourceName,
+				ImportStateIdFunc:       testAccCheckMongoDBAtlasProjectImportStateIDFunc(resourceName),
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"with_default_alerts_settings"},
+			},
+		},
+	})
+}
+
+func TestAccProjectRSProject_withUpdatedLimits(t *testing.T) {
+	var (
+		resourceName = "mongodbatlas_project.test"
+		projectName  = acctest.RandomWithPrefix("tf-acc-project")
+		orgID        = os.Getenv("MONGODB_ATLAS_ORG_ID")
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID, []*projectLimit{
+					{
+						name:  "atlas.project.deployment.clusters",
+						value: 1,
+					},
+					{
+						name:  "atlas.project.deployment.nodesPerPrivateLinkRegion",
+						value: 1,
+					},
+				}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "limits.0.name", "atlas.project.deployment.clusters"),
+					resource.TestCheckResourceAttr(resourceName, "limits.0.value", "1"),
+					resource.TestCheckResourceAttr(resourceName, "limits.1.name", "atlas.project.deployment.nodesPerPrivateLinkRegion"),
+					resource.TestCheckResourceAttr(resourceName, "limits.1.value", "1"),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID, []*projectLimit{
+					{
+						name:  "atlas.project.deployment.nodesPerPrivateLinkRegion",
+						value: 2,
+					},
+				}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+					resource.TestCheckResourceAttr(resourceName, "limits.0.name", "atlas.project.deployment.nodesPerPrivateLinkRegion"),
+					resource.TestCheckResourceAttr(resourceName, "limits.0.value", "2"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccProjectRSProject_withInvalidLimitName(t *testing.T) {
+	var (
+		projectName = acctest.RandomWithPrefix("tf-acc-project")
+		orgID       = os.Getenv("MONGODB_ATLAS_ORG_ID")
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID, []*projectLimit{
+					{
+						name:  "incorrect.name",
+						value: 1,
+					},
+				}),
+				ExpectError: regexp.MustCompile("Limit not found"),
+			},
+		},
+	})
+}
+
+func TestAccProjectRSProject_withInvalidLimitNameOnUpdate(t *testing.T) {
+	var (
+		resourceName = "mongodbatlas_project.test"
+		projectName  = acctest.RandomWithPrefix("tf-acc-project")
+		orgID        = os.Getenv("MONGODB_ATLAS_ORG_ID")
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheckBasic(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckMongoDBAtlasProjectDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID, []*projectLimit{}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", projectName),
+					resource.TestCheckResourceAttr(resourceName, "org_id", orgID),
+				),
+			},
+			{
+				Config: testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID, []*projectLimit{
+					{
+						name:  "incorrect.name",
+						value: 1,
+					},
+				}),
+				ExpectError: regexp.MustCompile("Limit not found"),
+			},
+		},
+	})
+}
+
+func testAccCheckMongoDBAtlasProjectExists(resourceName string, project *matlas.Project) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		conn := testMongoDBClient.(*MongoDBClient).Atlas
 
@@ -145,7 +424,7 @@ func testAccCheckMongoDBAtlasProjectExists2(resourceName string, project *matlas
 	}
 }
 
-func testAccCheckMongoDBAtlasProjectAttributes2(project *matlas.Project, projectName string) resource.TestCheckFunc {
+func testAccCheckMongoDBAtlasProjectAttributes(project *matlas.Project, projectName string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if project.Name != projectName {
 			return fmt.Errorf("bad project name: %s", project.Name)
@@ -153,4 +432,139 @@ func testAccCheckMongoDBAtlasProjectAttributes2(project *matlas.Project, project
 
 		return nil
 	}
+}
+
+func testAccCheckMongoDBAtlasProjectDestroy(s *terraform.State) error {
+	conn := testMongoDBClient.(*MongoDBClient).Atlas
+
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "mongodbatlas_project" {
+			continue
+		}
+
+		projectRes, _, _ := conn.Projects.GetOneProjectByName(context.Background(), rs.Primary.ID)
+		if projectRes != nil {
+			return fmt.Errorf("project (%s) still exists", rs.Primary.ID)
+		}
+	}
+
+	return nil
+}
+
+func testAccCheckMongoDBAtlasProjectImportStateIDFunc(resourceName string) resource.ImportStateIdFunc {
+	return func(s *terraform.State) (string, error) {
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return "", fmt.Errorf("not found: %s", resourceName)
+		}
+
+		return rs.Primary.ID, nil
+	}
+}
+
+func testAccMongoDBAtlasProjectConfig(projectName, orgID string, teams []*matlas.ProjectTeam) string {
+	var ts string
+
+	for _, t := range teams {
+		ts += fmt.Sprintf(`
+		teams {
+			team_id = "%s"
+			role_names = %s
+		}
+		`, t.TeamID, strings.ReplaceAll(fmt.Sprintf("%+q", t.RoleNames), " ", ","))
+	}
+
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name  			 = "%s"
+			org_id 			 = "%s"
+
+			%s
+		}
+	`, projectName, orgID, ts)
+}
+
+func testAccMongoDBAtlasProjectConfigWithUpdatedRole(projectName, orgID, teamID, roleName string) string {
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   = "%s"
+			org_id = "%s"
+
+			teams {
+				team_id = "%s"
+				role_names = ["%s"]
+			}
+		}
+	`, projectName, orgID, teamID, roleName)
+}
+
+func testAccMongoDBAtlasProjectConfigWithProjectOwner(projectName, orgID, projectOwnerID string) string {
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   			 = "%[1]s"
+			org_id 			 = "%[2]s"
+		    project_owner_id = "%[3]s"
+		}
+	`, projectName, orgID, projectOwnerID)
+}
+
+func testAccMongoDBAtlasGovProjectConfigWithProjectOwner(projectName, orgID, projectOwnerID string) string {
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   			 = "%[1]s"
+			org_id 			 = "%[2]s"
+		    project_owner_id = "%[3]s"
+			region_usage_restrictions = "GOV_REGIONS_ONLY"
+		}
+	`, projectName, orgID, projectOwnerID)
+}
+
+func testAccMongoDBAtlasProjectConfigWithFalseDefaultSettings(projectName, orgID, projectOwnerID string) string {
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   			 = "%[1]s"
+			org_id 			 = "%[2]s"
+			project_owner_id = "%[3]s"
+			with_default_alerts_settings = false
+		}
+	`, projectName, orgID, projectOwnerID)
+}
+
+func testAccMongoDBAtlasProjectConfigWithFalseDefaultAdvSettings(projectName, orgID, projectOwnerID string) string {
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   			 = "%[1]s"
+			org_id 			 = "%[2]s"
+			project_owner_id = "%[3]s"
+			with_default_alerts_settings = false
+			is_collect_database_specifics_statistics_enabled = false
+			is_data_explorer_enabled = false
+			is_extended_storage_sizes_enabled = false
+			is_performance_advisor_enabled = false
+			is_realtime_performance_panel_enabled = false
+			is_schema_advisor_enabled = false
+		}
+	`, projectName, orgID, projectOwnerID)
+}
+
+func testAccMongoDBAtlasProjectConfigWithLimits(projectName, orgID string, limits []*projectLimit) string {
+	var limitsString string
+
+	for _, limit := range limits {
+		limitsString += fmt.Sprintf(`
+		limits {
+			name = "%s"
+			value = %d
+		}
+		`, limit.name, limit.value)
+	}
+
+	return fmt.Sprintf(`
+		resource "mongodbatlas_project" "test" {
+			name   			 = "%s"
+			org_id 			 = "%s"
+
+			%s
+		}
+	`, projectName, orgID, limitsString)
 }
